refactor application folder structure
