+++ src/server/jestSetup.js
@@ -1,4 +1,4 @@
 const supertest = require('supertest');
-const app = require('../server');
+const app = require('./server');
 
 global.request = supertest(app);
